requirements.txt
chainlit
openai-agents
python-dotenv
requests